Extract components to seperate files
