+++ src/taxes/components/TaxChannelsMenu/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./TaxChannelsMenu";
+export * from "./TaxChannelsMenu";
