+++ src/taxes/components/TaxSettingsCard/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./TaxSettingsCard";
+export * from "./TaxSettingsCard";
